refactor(shader-knowladge): rewrite with public api version of shader information
