+++ Editor/APIInternal/ShaderInformation.Liltoon.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: a33c006ffa3441e98ed2d8482a9560f0
+timeCreated: 1726477905
